backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
python-multipart
gtts
jinja2
pyspellchecker
